#760 [REFACTOR] 메인페이지 Card, Community, Besookt 영역을 컴포넌트로 분리
