+++ src/pages/MainPage/MainPage.module.css
@@ -6,18 +6,6 @@
   margin-bottom: 2.5625rem;
 }
 
-.cards {
-  margin: 1.5rem var(--default-margin) 1.75rem var(--default-margin);
-}
-
-.notice {
-  flex: 2;
-}
-
-.mark {
-  flex: 3;
-}
-
 .community {
   margin-bottom: 25px;
 }
