+++ src/pages/MainPage/components/HomeBesookt/HomeBesookt.module.css
@@ -1,0 +1,9 @@
+.layout {
+  margin: 0 var(--default-margin);
+}
+
+.list {
+  display: flex;
+  flex-direction: column;
+  gap: 0.375rem;
+}
