+++ src/pages/MainPage/components/HomeCommunity/HomeCommunity.module.css
@@ -1,0 +1,8 @@
+.layout {
+  margin: 0 var(--default-margin);
+}
+
+.list {
+  display: flex;
+  gap: 0.625rem;
+}
