+++ src/pages/MainPage/components/index.js
@@ -1,0 +1,3 @@
+export { default as HomeCard } from './HomeCard/HomeCard';
+export { default as HomeCommunity } from './HomeCommunity/HomeCommunity';
+export { default as HomeBesookt } from './HomeBesookt/HomeBesookt';
